requirements.txt
python-dotenv
requests
cdsapi
netCDF4
geopandas
geojson